Change disk.offset to 64 bits to support files greater than 4gb
(cherry picked from commit ddb75455dfd2416ac3fa746c1c844669a20c4a33)

+++ include/nuster/disk.h
@@ -85,7 +85,7 @@ enum {
 typedef struct nst_disk_data {
     char               *file;               /* disk file */
     int                 fd;
-    int                 offset;
+    uint64_t            offset;
     char                meta[NST_DISK_META_SIZE];
 } nst_disk_data_t;
